alx_travel_app/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
celery>=5.3.0
python-dotenv>=1.0.0
requests>=2.31.0
# For RabbitMQ support
kombu>=5.3.0
# Email support
# If using a specific email backend, add those dependencies here

# Optional: For development
# django-debug-toolbar>=4.2.0
